+++ R/horns-plot-for-export.R
@@ -1,23 +1,15 @@
 #' Visualize horns index (\eqn{h}) frequencies
 #'
-#' @description Two functions that visualize the distribution of horns index
-#'   values found by CLOSURE:
+#' @description `closure_plot_horns_histogram()` visualizes the distribution of
+#'   horns index values found by CLOSURE.
 #'
-#'   - `closure_plot_horns_histogram()` bins the distribution into categories,
-#'   emphasizing individual groups of values. You can adjust this via the
-#'   `binwidth` argument.
-#'   - `closure_plot_horns_density()` smooths the distribution, emphasizing its
-#'   shape. You can use the `density_bounds` argument to avoid apparent density
-#'   beyond the limits.
-#'
-#'   Both mark the minimum and maximum values. The x-axis always ranges from 0
-#'   to 1. This reveals the big picture, putting any variability among horns
-#'   values into perspective.
+#'   It marks the min, max, and uniform values; see [`horns_uniform()`]. The
+#'   x-axis always ranges from 0 to 1. This reveals the big picture, putting any
+#'   variability among horns values into perspective.
 #'
 #' @inheritParams closure_plot_bar
-#' @param alpha Numeric (length 1). Opacity of the density or bars. Default is
-#'   `0.75`.
-#' @param color String (length 1). Color of the density or bars. Default is
+#' @param alpha Numeric (length 1). Opacity of the bars. Default is `0.75`.
+#' @param bar_color String (length 1). Color of the bars. Default is
 #'   `"#5D3FD3"`, a purple color.
 #' @param binwidth Numeric (length 1). Only in `closure_plot_horns_histogram()`.
 #'   Width of the bins that divide up the x-axis. Default is `0.01`.
@@ -28,11 +20,6 @@
 #'   - `"min_max"`, `"min_max_uniform"`, `"min_max_bounds"`, `"uniform"`,
 #'   `"uniform_bounds"`, or `"bounds"`: a subset of these labels.
 #'   - `"none"`: no labels.
-#' @param density_bounds String (length 1). Only in
-#'   `closure_plot_horns_density()`. Which limits, if any, should the density be
-#'   forced to fit between? Default is `"none"`. If set to `"min_max"`, this
-#'   avoids the illusion of points beyond the limits but can lead to a U-shaped
-#'   effect, which would also be misleading. Default is `0.01`.
 #' @param line_color_min_max Numeric (length 1). Color of the lines that mark
 #'   the lower and upper ends of the distribution. Default is `"red"`.
 #' @param text_limits Numeric (length 2). If the minimum horns index is lower
@@ -57,8 +44,6 @@
 #'
 #' closure_plot_horns_histogram(data_near_zero)
 #'
-#' closure_plot_horns_density(data_near_zero)
-#'
 #' data_near_midpoint <- closure_generate(
 #'   mean = "2.8",
 #'   sd = "1.5",
@@ -68,8 +53,6 @@
 #' )
 #'
 #' closure_plot_horns_histogram(data_near_midpoint)
-#'
-#' closure_plot_horns_density(data_near_midpoint)
 #'
 #' # Large difference between horns values occur (only?)
 #' # if there are no decimal places in `mean` and `sd`:
@@ -82,19 +65,8 @@
 #' )
 #'
 #' closure_plot_horns_histogram(data_wide_spread)
-#'
-#' closure_plot_horns_density(data_wide_spread)
 
 closure_plot_horns_histogram <- new_plot_fn_horns_frequency(
   technique = "CLOSURE",
-  type = "histogram"
-)
-
-
-#' @rdname horns-frequency
-#' @export
-
-closure_plot_horns_density <- new_plot_fn_horns_frequency(
-  technique = "CLOSURE",
-  type = "density"
+  bar_color = "#5D3FD3"
 )